# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Microservice for antivirus scanning",
        "title": "Sereni Antivirus API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8084",
    "basePath": "/",
    "paths": {
        "/scan": {
            "post": {
                "description": "Upload a file to scan it for malware using ClamAV",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "antivirus"
                ],
                "summary": "Scan a file for viruses",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to scan",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/interfaces.ScanResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/scan-files": {
            "post": {
                "description": "Upload multiple files to scan them for malware",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "antivirus"
                ],
                "summary": "Scan multiple files for viruses",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "csv",
                        "description": "Files to scan",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/interfaces.ScanResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "interfaces.ScanResult": {
            "type": "object",
            "properties": {
                "clean": {
                    "description": "Clean indicates whether the content was found to be clean",
                    "type": "boolean"
                },
                "file_name": {
                    "type": "string"
                },
                "threat": {
                    "description": "Threat describes the detected threat, if any. Empty when Clean is true",
                    "type": "string"
                }
            }
        }
    }
}